requirements.txt
https://github.com/happyleavesaoc/spotipy/archive/544614f4b1d508201d363e84e871f86c90aa26b2.zip#spotipy==2.4.4
fuzzywuzzy==0.15.1
dateparser==0.6.0
python-forecastio==1.3.5
